Reorder project files
