+++ pkg/system/system.go
@@ -1,0 +1,3 @@
+package system
+
+// Keybindings + system config
