--- src/system/system.go
@@ -1,3 +1,0 @@
-package system
-
-// Keybindings + system config
